requirements.txt
Flask==0.10.1
Flask-PyMongo==0.3.0
Flask-RESTful==0.3.5
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
aniso8601==0.82
gunicorn==19.0.0
itsdangerous==0.24
pymongo==2.8
pytz==2014.4
six==1.7.2
readability-api=1.0.2